# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Swapl API
  version: 0.1.0
  description: >-
    Shared contract for the Swapl home-swap API consumed by the web (Next.js),
    iOS, and Android clients. Server-authoritative: every platform calls this
    one API, which owns the single Postgres database. Schemas mirror the
    client DTOs proven to decode the live responses today.
servers:
  - url: http://localhost:3000
    description: Local dev
  - url: https://swapl.fun
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Opaque token from POST /api/auth/token (native clients).
    cookieAuth:
      type: apiKey
      in: cookie
      name: swapl_session
      description: Signed session cookie (web).
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
        alreadyVerified: { type: boolean }
...
paths:
...
  /api/agreements/{id}/cancel:
    post:
      tags: [agreements]
      summary: Cancel a swap before it starts
      description: >-
        Marks an ACTIVE agreement INTERRUPTED, cancels the insurance policy
        at the underwriter and notifies both parties. Only before dateFrom —
        mid-stay interruptions go through the claim flow instead.
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Cancelled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
        '401': { description: Unauthenticated }
        '403': { description: Not a party of the agreement }
        '404': { description: Not found }
        '409': { description: Already closed, or the cancellation window has passed }
...
  /api/concierge/checkout:
    post:
      tags: [concierge]
      summary: Buy a concierge add-on for a swap
      description: >-
        Stripe Checkout when configured (`{ url, mode: "checkout" }`),
        otherwise records the order immediately
        (`{ ok, mode: "pre_launch", orderId }`). Affiliate add-ons are
        rejected — they go through the affiliate redirector.
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agreementId, slug]
              properties:
                agreementId: { type: string }
                slug: { type: string }
      responses:
        '200':
          description: Checkout URL or recorded pre-launch order
          content:
            application/json:
              schema:
                type: object
                required: [mode]
                properties:
                  mode: { type: string, enum: [checkout, pre_launch] }
                  url: { type: string }
                  ok: { type: boolean }
                  orderId: { type: string }
        '400': { description: Invalid input, or affiliate add-on }
        '401': { description: Unauthenticated }
        '403': { description: Not a party of the agreement }
        '404': { description: Agreement or add-on not found }
...
  /api/corporate/leads:
    post:
      tags: [corporate]
      summary: Submit a corporate lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [companyName, contactName, email]
              properties:
                companyName: { type: string, minLength: 2 }
                contactName: { type: string, minLength: 2 }
                email: { type: string, format: email }
                phone: { type: string }
                employeeCount: { type: integer, minimum: 1 }
                useCase: { type: string, maxLength: 2000 }
      responses:
        '200':
          description: Lead recorded
          content:
            application/json:
              schema:
                type: object
                required: [ok, id]
                properties:
                  ok: { type: boolean }
                  id: { type: string }
        '400': { description: Invalid input }
  /api/listings/featured:
    post:
      tags: [listings]
      summary: Buy a Featured Placement for my listing
      description: >-
        14 or 30 days. Stripe Checkout when configured
        (`{ url, mode: "checkout" }`), otherwise activates immediately
        (`{ ok, mode: "pre_launch", endsAt }`).
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [listingId, durationDays]
              properties:
                listingId: { type: string }
                durationDays: { type: integer, enum: [14, 30] }
      responses:
        '200':
          description: Checkout URL or immediate activation
          content:
            application/json:
              schema:
                type: object
                required: [mode]
                properties:
                  mode: { type: string, enum: [checkout, pre_launch] }
                  url: { type: string }
                  ok: { type: boolean }
                  endsAt: { type: string, format: date-time }
        '400': { description: Invalid input }
        '401': { description: Unauthenticated }
        '404': { description: Listing not found or not mine }
  /api/listings/verify:
    post:
      tags: [listings]
      summary: Submit my listing for paid verification
      description: >-
        Stripe Checkout when configured (`{ url, mode: "checkout" }`),
        otherwise records the pending submission directly
        (`{ ok, mode: "pre_launch" }`) for the admin queue.
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [listingId, videoUrl]
              properties:
                listingId: { type: string }
                videoUrl: { type: string, format: uri }
      responses:
        '200':
          description: Checkout URL or recorded submission
          content:
            application/json:
              schema:
                type: object
                required: [mode]
                properties:
                  mode: { type: string, enum: [checkout, pre_launch] }
                  url: { type: string }
                  ok: { type: boolean }
        '400': { description: Invalid input }
        '401': { description: Unauthenticated }
        '404': { description: Listing not found or not mine }
        '409': { description: Already submitted or verified }
...
  /api/marketing/events:
    post:
      tags: [marketing]
      summary: Record a marketing event
      description: Anonymous first-party attribution event (page, UTM fields, optional metadata). Public.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eventName]
              properties:
                eventName: { type: string, maxLength: 80 }
                path: { type: string, maxLength: 500 }
                source: { type: string, maxLength: 80 }
                medium: { type: string, maxLength: 80 }
                campaign: { type: string, maxLength: 120 }
                term: { type: string, maxLength: 120 }
                content: { type: string, maxLength: 120 }
                referrer: { type: string, maxLength: 500 }
                metadata: { type: object, additionalProperties: true }
      responses:
        '200':
          description: Recorded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
        '400': { description: Invalid input }
...